In this version, the already processed data will be skip instead of processing again in the previous version

+++ ExtractGE.py
@@ -16,11 +16,12 @@ info = "The time spend on processing {0} is: {1:.6f}s"
 for sub_dir in gs_dir.iterdir():
     Path(str(sub_dir).replace("GS", "GE")).mkdir(parents=True, exist_ok=True)
     for gs_file in sub_dir.iterdir():
-        t0 = time()
-        ge_file = str(gs_file).replace("GS", "GE").replace("npz", "npy")
-        with np.load(gs_file) as ld:
-            alpha, beta = ld["parameters"]
-            gse = ld["gse"][0]
-        np.save(ge_file, [alpha, beta, gse])
-        t1 = time()
-        print(info.format(gs_file.name, t1 - t0))
+        ge_file = Path(str(gs_file).replace("GS", "GE").replace("npz", "npy"))
+        if not ge_file.exists():
+            t0 = time()
+            with np.load(gs_file) as ld:
+                alpha, beta = ld["parameters"]
+                gse = ld["gse"][0]
+            np.save(ge_file, [alpha, beta, gse])
+            t1 = time()
+            print(info.format(gs_file.name, t1 - t0))
